requirements.txt
airfrans==0.1.5.1
datasets==4.1.1
hydra-core==1.3.2
matplotlib==3.8.4
numpy==2.3.3
omegaconf==2.3.0
Pillow==11.3.0
plaid==0.1
point_cloud_utils==0.30.4
pyoche==0.1.3
PyYAML==6.0.1
PyYAML==6.0.3
scikit_learn==1.4.2
scipy==1.16.2
torch==2.2.2+cu118
torch_geometric==2.6.1
tqdm==4.66.2
